src/prediction_pipeline/requirements.txt
fastapi
uvicorn==0.17.6
joblib~=1.0
numpy~=1.20
google-cloud-storage>=1.26.0,<2.0.0dev
google-cloud-aiplatform[prediction]>=1.16.0
beautifulsoup4==4.11.1
transformers==4.25.1
torch==1.13.1